Revert "♻️ Refactor chat message (3)"
This reverts commit b1ad563b1a008a85b76af76938b87458e54d7da3.
