+++ lib/widgets/widgets.dart
@@ -7,5 +7,3 @@ export 'requireEmailUpfront.dart';
 export 'alert.dart';
 export 'timeWidget.dart';
 export 'dayIndicator.dart';
-export 'mdViewer.dart';
-export 'userAvatar.dart';
